+++ theicenet-cryptography-module/src/test/java/com/theicenet/cryptography/cipher/asymmetric/rsa/JCARSACipherServiceTest.java
@@ -15,6 +15,8 @@
  */
 package com.theicenet.cryptography.cipher.asymmetric.rsa;
 
+import static com.theicenet.cryptography.test.support.KeyPairUtil.toPrivateKey;
+import static com.theicenet.cryptography.test.support.KeyPairUtil.toPublicKey;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.collection.IsCollectionWithSize.hasSize;
 import static org.hamcrest.core.Is.is;
@@ -29,9 +31,7 @@ import java.nio.charset.StandardCharsets;
 import java.security.KeyFactory;
 import java.security.PrivateKey;
 import java.security.PublicKey;
-import java.security.spec.PKCS8EncodedKeySpec;
 import java.security.spec.RSAPrivateKeySpec;
-import java.security.spec.X509EncodedKeySpec;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
@@ -101,8 +101,8 @@ class JCARSACipherServiceTest {
               + "5665d74c109c84e02ef91191fd3561f3363cd2a5d9dbc622dfb35e4fa9b9a5bce"
               + "0f4d3349d127f4902bfc2d8b1e980");
 
-  final PublicKey RSA_PUBLIC_KEY_2048_BITS;
-  final PrivateKey RSA_PRIVATE_KEY_2048_BITS;
+  final PublicKey RSA_PUBLIC_KEY_2048_BITS = toPublicKey(RSA_PUBLIC_KEY_2048_BITS_BYTE_ARRAY, RSA);
+  final PrivateKey RSA_PRIVATE_KEY_2048_BITS = toPrivateKey(RSA_PRIVATE_KEY_2048_BITS_BYTE_ARRAY, RSA);
 
   final byte[] RSA_ENCRYPTED_OAEP_WITH_SHA1_AND_MGF1_PADDING =
       HexUtil.decodeHex(
@@ -114,16 +114,6 @@ class JCARSACipherServiceTest {
               + "aba310fe72bbc5ff15ed7f848490a3b9a0877152697abf4b1754f4f6a7f09d96f729d6d"
               + "a5ade53bc297e6e6c218c379d6d99ab5278c68137d83006f3f7b2921672de3670f1055d"
               + "8e7400222");
-
-  JCARSACipherServiceTest() throws Exception {
-    final var keyFactory = KeyFactory.getInstance(RSA);
-
-    final var x509EncodedKeySpec = new X509EncodedKeySpec(RSA_PUBLIC_KEY_2048_BITS_BYTE_ARRAY);
-    RSA_PUBLIC_KEY_2048_BITS = keyFactory.generatePublic(x509EncodedKeySpec);
-
-    final var pkcs8EncodedKeySpec = new PKCS8EncodedKeySpec(RSA_PRIVATE_KEY_2048_BITS_BYTE_ARRAY);
-    RSA_PRIVATE_KEY_2048_BITS = keyFactory.generatePrivate(pkcs8EncodedKeySpec);
-  }
 
   @ParameterizedTest
   @EnumSource(RSAPadding.class)