+++ theicenet-cryptography-module/src/test/java/com/theicenet/cryptography/keyagreement/ecc/ecdh/JCACEDHKeyAgreementServiceTest.java
@@ -15,6 +15,8 @@
  */
 package com.theicenet.cryptography.keyagreement.ecc.ecdh;
 
+import static com.theicenet.cryptography.test.support.KeyPairUtil.toPrivateKey;
+import static com.theicenet.cryptography.test.support.KeyPairUtil.toPublicKey;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.collection.IsCollectionWithSize.hasSize;
 import static org.hamcrest.core.Is.is;
@@ -22,23 +24,26 @@ import static org.hamcrest.core.IsEqual.equalTo;
 import static org.hamcrest.core.IsNull.notNullValue;
 import static org.hamcrest.number.OrderingComparison.greaterThan;
 
-import com.theicenet.cryptography.key.asymmetric.ecc.ECCKeyAlgorithm;
 import com.theicenet.cryptography.keyagreement.KeyAgreementService;
 import com.theicenet.cryptography.test.support.HexUtil;
 import com.theicenet.cryptography.test.support.RunnerUtil;
 import com.theicenet.cryptography.util.CryptographyProviderUtil;
-import java.security.KeyFactory;
 import java.security.PrivateKey;
 import java.security.PublicKey;
-import java.security.spec.PKCS8EncodedKeySpec;
-import java.security.spec.X509EncodedKeySpec;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 /**
  * @author Juan Fidalgo
  */
 class JCACEDHKeyAgreementServiceTest {
-  final ECCKeyAlgorithm ECDH = ECCKeyAlgorithm.ECDH;
+
+  static {
+    // Bouncy Castle is required to reformat the ECDH public and private keys
+    CryptographyProviderUtil.addBouncyCastleCryptographyProvider();
+  }
+
+  final String ECDH = "ECDH";
 
   final byte[] ECDH_PUBLIC_KEY_BRAINPOOLP256R1_BYTE_ARRAY_ALICE =
       HexUtil.decodeHex(
@@ -66,38 +71,22 @@ class JCACEDHKeyAgreementServiceTest {
               + "47e0175135e72050c56fb9c3a97db56370123b66e5ebec702bcc5889149628822b169c9678"
               + "30499668d78eb5f38ec437eef1c8dab3fac2896ec6b5c0f534");
 
-  final PublicKey ECDH_PUBLIC_KEY_BRAINPOOLP256R1_ALICE;
-  final PrivateKey ECDH_PRIVATE_KEY_BRAINPOOLP256R1_ALICE;
-  final PublicKey ECDH_PUBLIC_KEY_BRAINPOOLP256R1_BOB;
-  final PrivateKey ECDH_PRIVATE_KEY_BRAINPOOLP256R1_BOB;
+  final PublicKey ECDH_PUBLIC_KEY_BRAINPOOLP256R1_ALICE =
+      toPublicKey(ECDH_PUBLIC_KEY_BRAINPOOLP256R1_BYTE_ARRAY_ALICE, ECDH);
+  final PrivateKey ECDH_PRIVATE_KEY_BRAINPOOLP256R1_ALICE =
+      toPrivateKey(ECDH_PRIVATE_KEY_BRAINPOOLP256R1_BYTE_ARRAY_ALICE, ECDH);
+  final PublicKey ECDH_PUBLIC_KEY_BRAINPOOLP256R1_BOB =
+      toPublicKey(ECDH_PUBLIC_KEY_BRAINPOOLP256R1_BYTE_ARRAY_BOB, ECDH);
+  final PrivateKey ECDH_PRIVATE_KEY_BRAINPOOLP256R1_BOB =
+      toPrivateKey(ECDH_PRIVATE_KEY_BRAINPOOLP256R1_BYTE_ARRAY_BOB, ECDH);
 
   final byte[] ECDH_DERIVED_SECRET_KEY =
       HexUtil.decodeHex("3078620e26babfd1200f70a280f7370ef15ce0176e983a2f6803de6eff5dc269");
 
-  final KeyAgreementService keyAgreementService;
+  KeyAgreementService keyAgreementService;
 
-  JCACEDHKeyAgreementServiceTest() throws Exception {
-    // Bouncy Castle is required for ECDH key factory
-    CryptographyProviderUtil.addBouncyCastleCryptographyProvider();
-
-    final var keyFactory = KeyFactory.getInstance(ECDH.toString());
-
-    final var x509EncodedKeySpecAlice = new X509EncodedKeySpec(
-        ECDH_PUBLIC_KEY_BRAINPOOLP256R1_BYTE_ARRAY_ALICE);
-    ECDH_PUBLIC_KEY_BRAINPOOLP256R1_ALICE = keyFactory.generatePublic(x509EncodedKeySpecAlice);
-
-    final var pkcs8EncodedKeySpecAlice = new PKCS8EncodedKeySpec(
-        ECDH_PRIVATE_KEY_BRAINPOOLP256R1_BYTE_ARRAY_ALICE);
-    ECDH_PRIVATE_KEY_BRAINPOOLP256R1_ALICE = keyFactory.generatePrivate(pkcs8EncodedKeySpecAlice);
-
-    final var x509EncodedKeySpecBob = new X509EncodedKeySpec(
-        ECDH_PUBLIC_KEY_BRAINPOOLP256R1_BYTE_ARRAY_BOB);
-    ECDH_PUBLIC_KEY_BRAINPOOLP256R1_BOB = keyFactory.generatePublic(x509EncodedKeySpecBob);
-
-    final var pkcs8EncodedKeySpecBob = new PKCS8EncodedKeySpec(
-        ECDH_PRIVATE_KEY_BRAINPOOLP256R1_BYTE_ARRAY_BOB);
-    ECDH_PRIVATE_KEY_BRAINPOOLP256R1_BOB = keyFactory.generatePrivate(pkcs8EncodedKeySpecBob);
-
+  @BeforeEach
+  void setUp() {
     keyAgreementService = new JCACEDHKeyAgreementService();
   }